relocate ResponseMode enum
