+++ apps/app/src/server/service/file-uploader/aws.ts
@@ -11,14 +11,14 @@ import {
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
 import urljoin from 'url-join';
 
-import type { RespondOptions } from '~/server/interfaces/attachment';
+import { ResponseMode, type RespondOptions } from '~/server/interfaces/attachment';
 import type { IAttachmentDocument } from '~/server/models';
 import loggerFactory from '~/utils/logger';
 
 import { configManager } from '../config-manager';
 
 import {
-  AbstractFileUploader, ResponseMode, type TemporaryUrl, type SaveFileParam,
+  AbstractFileUploader, type TemporaryUrl, type SaveFileParam,
 } from './file-uploader';
 import { ContentHeaders } from './utils';
