+++ apps/app/src/server/service/file-uploader/azure.ts
@@ -16,14 +16,14 @@ import {
   type BlockBlobUploadStreamOptions,
 } from '@azure/storage-blob';
 
-import type { RespondOptions } from '~/server/interfaces/attachment';
+import { ResponseMode, type RespondOptions } from '~/server/interfaces/attachment';
 import type { IAttachmentDocument } from '~/server/models';
 import loggerFactory from '~/utils/logger';
 
 import { configManager } from '../config-manager';
 
 import {
-  AbstractFileUploader, ResponseMode, type TemporaryUrl, type SaveFileParam,
+  AbstractFileUploader, type TemporaryUrl, type SaveFileParam,
 } from './file-uploader';
 import { ContentHeaders } from './utils';
